Fix gesture method signatures
