+++ .github/workflows/build_apk.yml
@@ -6,6 +6,8 @@ on:
 jobs:
   build:
     runs-on: ubuntu-latest
+    container:
+      image: cirrusci/android-sdk:33
     steps:
       - uses: actions/checkout@v3
       - name: Setup JDK 17
@@ -13,11 +15,7 @@ jobs:
         with:
           distribution: temurin
           java-version: '17'
-      - name: Setup Android SDK
-        uses: android-actions/setup-android@v3
-        with:
-          packages: 'platforms;android-34 build-tools;34.0.0'
-      - run: echo "sdk.dir=$ANDROID_SDK_ROOT" > local.properties
+      - run: echo "sdk.dir=$ANDROID_HOME" > local.properties
       - name: Grant execute permission for gradlew
         run: chmod +x gradlew
       - name: Build APK
